[native pos] Detect CPP process crash and fail native task
Currently, the executor JVM waits on the remote task either
reaching a terminal state or returning results. Hence, when
the cpp process crashes, neither of these conditions are fulfilled
and process JVM hangs forever.

Also, this situation is already dt etected by the JVM, through the taskInfo
error tracker.

This PR propagates this information on the caller/waiter so that we can
fail the task instead of hanging