[native pos] Fix TaskInfo fetcher logic to avoid concurrent runs
Currently, the TaskInfoFetcher schedules an async call to
fetch TaskInfo. But the call could take a long time or timeout
A new instance could start before that happens and we would
have undesirable behavior.

The scheduleAtFixedDelay allows us to guard against this by ensuring
that we do not start the next instance of the work (http call) before
the first one returns. But because we add a async callback in that thread
it returns early and disables this mechanism.

This commit makes the logic inside of the scheduled thread to sync so that
that we can leverage the thread scheduler to guard us from synchronous calls