[native pos] Allow additive updates to spark accumulators
Current implementation of accumulator updates extract values
from taskInfo that have the current value of the metric instead
of the incremental/diff value to be 'add'ed to the spark accum
This makes the util unsuable if we want to continuously update
the metrics. This continuous updates are needed to ensure
spark sees the tasks progress through (inputBytesRead, shuffleBytesRead)
and can kill the task if it isn't making progress.

This commit changes the implementation to compute the diff between
accum current value and new value in taskInfo and invoke accum.add