[pos][native] Reduce testing log bloat to enable better CI failures debugging
Current logging of presto-on-spark leads to 100's of Mb of log files in CI.
This results in being unable to use github UI to see the error info.

Mainly the logs are being flooded by 2 classes PrestoSparkTaskExecutorFactory and spark's TaskSetManager.
In this PR, we change the log level to WARN to omit those logs